requirements.txt
flask==3.0.0
flask-pymongo==2.3.0
pymongo==4.6.1
flask-login==0.6.3
flask-bcrypt==1.0.1
python-dotenv==1.0.0
dnspython==2.4.2
